Fix buf size when receiving ANSI escapes via UART
Supported ANSI escapes consist of SCI "\e[" followed by:
1. A optional number (max 3 digits) specified in ASCII followed by a
   control character.
2. A optional number (max 3 digits) followed by ';' and then by another
   optional number (max 3 digits), concluded with a control character.

This means that when receiving char by char via UART a buffer needs to
be allocated that can store the longest possible ANSI escape sequence.
This would be:

 2  +      3       + 1 +     3         +        1          = 10
 |         |         |       |                  |
CSI   First number   ;   Second number   Control Character

Note however the buffer needs to be null character terminated, thus the
buffer must be 11 bytes. In the future, an optimization could possibly
be achieved here by using malloc to do dynamic allocation as more memory
is needed instead of allocating for the worst case scenario to the stack
as is done here.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>